Split out invalid-xpaths code
